[IMP] runbot: increase cpu_limit
When the Odoo instances are spawned for tests, they have a time limit
set on their CPU usage. This limit is hard-coded in the _spawn method
calls.

As the number of tests are increasing, their duration increases too.
As this limit is inherited by subprocesses, if a phantomjs test
last too long, the test is killed alone.

Finally, when the coverage is enabled, the tests duration is
approximately increases of 1.5 times.

With this commit, the cpu_limit of the two main tests jobs are
increased. An ir.config_parameter can optionaly be used to tweak this
limit without changing the code. When the coverage is used, the
cpu_limit is increased.